separate logic and ui into 2 files (custom hook)
