+++ src/_auth/form/SignUpForm.tsx
@@ -1,8 +1,3 @@
-import { zodResolver } from "@hookform/resolvers/zod";
-import { useForm } from "react-hook-form";
-import * as z from "zod";
-
-import { useToast } from "@/components/ui/use-toast";
 import { Button } from "@/components/ui/button";
 import {
   Form,
@@ -13,64 +8,12 @@ import {
   FormMessage,
 } from "@/components/ui/form";
 import { Input } from "@/components/ui/input";
-import { SignUpValidation } from "@/lib/validation";
 import Loader from "@/components/ui/shared/Loader";
-import { Link, useNavigate } from "react-router-dom";
-import {
-  useCreateUserAccount,
-  useSignInAccount,
-} from "@/lib/react-query/QueriesAndMutaions";
-import { useUserContext } from "@/context/AuthContext";
+import useSignUpForm from "@/hooks/useSignUpForm";
+import { Link } from "react-router-dom";
 
 const SignUpForm = () => {
-  const { toast } = useToast();
-  const { mutateAsync: createUserAccount, isPending: isCreatingAccount } =
-    useCreateUserAccount();
-  const { checkAuthUser, isLoading: isUserLoading } = useUserContext();
-  const navigate = useNavigate();
-  const { mutateAsync: signInAccount } = useSignInAccount();
-
-  // 1. Define your form.
-  const form = useForm<z.infer<typeof SignUpValidation>>({
-    resolver: zodResolver(SignUpValidation),
-    defaultValues: {
-      name: "",
-      username: "",
-      email: "",
-      password: "",
-    },
-  });
-
-  // 2. Define a submit handler.
-  async function onSubmit(values: z.infer<typeof SignUpValidation>) {
-    const newUser = await createUserAccount(values);
-    if (!newUser) {
-      return toast({
-        title: "Sign up failed. Please try again",
-      });
-    }
-    const session = await signInAccount({
-      email: values.email,
-      password: values.password,
-    });
-
-    if (!session) {
-      return toast({
-        title: "Sign in failed. Please try again",
-      });
-    }
-
-    const isLoggedIn = await checkAuthUser();
-
-    if (isLoggedIn) {
-      form.reset();
-      navigate("/");
-    } else {
-      return toast({
-        title: "Sign in failed. Please try again",
-      });
-    }
-  }
+  const { form, isCreatingAccount, onSubmit, isUserLoading } = useSignUpForm();
 
   return (
     <Form {...form}>
